# server/infrastructure/context/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: COMATCH product development
  description: App to collect product development operational metrics
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /productDevelopments/{id}:
    get:
      tags:
        - ProductDevelopment
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the product development project to retrieve
          type: integer
      responses:
        200:
          description: Return the product development project with the specified id
        404:
          description: Example not found
...